Import contents for v2 of the site
This new version of the site has separate pages for each of the major
topics I want to cover.  To simplify the pain in keeping the pages in
sync with each other, I am now using the Jekyll static content
management system.

The contents of the site are roughly the same as before but they have changed
in various places in subtle ways.  There is no easy way to see what changed in
this diff due to the major overhaul of the file layout.