Document my username
The previous site had 'jmmv' in a very prominent place (the very top of
the front page), but this new version does not.  I have tried adding it
there but I do not like the looks of it.  Therefore, just put this
information in the about page with some commentary.