Move sections up one level
The fact that the title of the page is shown as h1 does not mean that the
rest of the sections should be h2 or higher.  In fact, we should treat the
title of the page as if it said "Introduction" and let the rest of the
article define sections at the same level.