Clean and homogenize page headers
Drop the cover image from the front page in favor of a simplified gradient
and consistently use the page-header class to mark all page headers.

As part of this, drop the "content" class hack that I used to properly
pad the top of the posts because my previous re-use of a header element
broke things.  The contents are now simpler.