Allow collapsing the front-page marketing buttons
As part of this change, introduce a dependency on the js-cookie library
and import the jquery library into our own tree.  For the former, this is
the recommended way of using this library and, for the latter, keeping
the sources in our tree will let us do local development without external
connections.