Change post layout to match Blogger
In preparation for the import of all posts from The Julipedia, change the
permalink layout to match the one from Blogger.

While doing so, add placeholder redirects from all posts that previously
existed on this site.  Because the number of original posts published here
was extremely limited (the majority were posts imported from Medium), and
because the existing posts have barely no incoming traffic, maintaining
this set of redirects by hand is simpler than anything else.