Convert The Julipedia posts to Jekyll posts
Run the import-blogger script over the external backup of The Julipedia and
add all resulting posts and drafts.

The posts are currently a verbatim copy of what was found in the Blogger
site: i.e. an HTML dump of their contents.  This does not fit well with the
style of the site, but is good enough for now.