Remove the base layout
In order to allow full customization of the layouts, which we will need to
use the jekyll-archives plugin, remove the base layout.  Instead, make each
layout file define the whole HTML page, and only include chunks for common
stuff.

As a side-effect, make the head.html include file NOT define the 'head'
element by itself so that the caller can add its own properties.  This will
help the above as well, and also cleans up the mess that was introduced
with the redirect.html layout.