Generate the sidebar archive out of band
The Liquid code previously introduced to generate the blog's archive was
quite complicated and insanely slow.  Replace it with some shell logic to
generate the archive contents out of band, which is much faster.

The downside is that Jekyll loses sight of the contents of the archive and
cannot regenerate it on the fly.  This is not a big deal though because it
will only affect the dev environment: for production, we always regenerate
everything from scratch.

See the comments in the newly-added generate-archive.sh script for details.

As a side-effect, add collapsing features to the archive.