Migrate compat post locations to aliases
When this static blog was originally created, I moved a bunch of posts
to it but not all.  Later on, when importing all of Blogger contents,
the layout of the posts had to change to avoid breaking those links.

For simplicity, I just added backwards-compatibility links for those
few posts.  Those were static pages, but with Hugo, they can simply
be converted into aliases.